refactor: convert CSS modules to Tailwind CSS
