+++ frontend/src/components/FileList/DragZone.tsx
@@ -1,14 +1,15 @@
 import React, { useRef, useState } from 'react';
-import styles from './DragZone.module.css';
+import { cn } from '../../lib/utils';
+import { Upload } from 'lucide-react';
 
 interface DragZoneProps {
-  onFileSelect: (files: FileList) => void;
+  onFilesDrop: (files: FileList) => void;
   multiple?: boolean;
   accept?: string;
 }
 
 const DragZone: React.FC<DragZoneProps> = ({
-  onFileSelect,
+  onFilesDrop,
   multiple = true,
   accept
 }) => {
@@ -39,7 +40,7 @@ const DragZone: React.FC<DragZoneProps> = ({
 
     const files = e.dataTransfer.files;
     if (files && files.length > 0) {
-      onFileSelect(files);
+      onFilesDrop(files);
     }
   };
 
@@ -50,13 +51,16 @@ const DragZone: React.FC<DragZoneProps> = ({
   const handleFileInput = (e: React.ChangeEvent<HTMLInputElement>) => {
     const files = e.target.files;
     if (files && files.length > 0) {
-      onFileSelect(files);
+      onFilesDrop(files);
     }
   };
 
   return (
     <div
-      className={`${styles.dragZone} ${isDragging ? styles.dragging : ''}`}
+      className={cn(
+        "border-2 border-dashed rounded-lg p-6 text-center bg-gray-50 cursor-pointer transition-all",
+        isDragging ? "border-blue-500 bg-blue-50" : "border-gray-300 hover:border-blue-500 hover:bg-gray-100"
+      )}
       onDragEnter={handleDragEnter}
       onDragLeave={handleDragLeave}
       onDragOver={handleDragOver}
@@ -66,16 +70,18 @@ const DragZone: React.FC<DragZoneProps> = ({
       <input
         type="file"
         ref={fileInputRef}
-        style={{ display: 'none' }}
+        className="hidden"
         onChange={handleFileInput}
         multiple={multiple}
         accept={accept}
       />
-      <div className={styles.content}>
-        <div className={styles.icon}>📤</div>
-        <div className={styles.text}>
-          <p>点击或拖拽文件到此处上传</p>
-          <p className={styles.hint}>支持单个或多个文件</p>
+      <div className="flex flex-col items-center gap-4">
+        <div className="text-blue-500">
+          <Upload size={32} />
+        </div>
+        <div className="flex flex-col gap-1">
+          <p className="text-gray-900 m-0">点击或拖拽文件到此处上传</p>
+          <p className="text-xs text-gray-600 m-0">支持单个或多个文件</p>
         </div>
       </div>
     </div>
@@ -83,3 +89,4 @@ const DragZone: React.FC<DragZoneProps> = ({
 };
 
 export default DragZone;
+
