--- frontend/src/pages/About/About.module.css
@@ -1,275 +1,0 @@
-.container {
-  max-width: 64rem;
-  margin: 0 auto;
-  padding: 2rem 1rem;
-}
-
-.header {
-  text-align: center;
-  margin-bottom: 3rem;
-}
-
-.title {
-  font-size: 2.25rem;
-  font-weight: bold;
-  color: #1f2937;
-  margin-bottom: 1rem;
-}
-
-.subtitle {
-  font-size: 1.125rem;
-  color: #6b7280;
-  max-width: 36rem;
-  margin: 0 auto;
-}
-
-.headerButtons {
-  display: flex;
-  justify-content: center;
-  gap: 1rem;
-  margin-top: 2rem;
-  margin-bottom: 2rem;
-}
-
-.githubLink {
-  display: inline-flex;
-  align-items: center;
-  gap: 8px;
-  color: #24292e;
-  background-color: #ffffff;
-  border: 1px solid #e1e4e8;
-  border-radius: 6px;
-  padding: 10px 16px;
-  font-size: 14px;
-  font-weight: 600;
-  text-decoration: none;
-  transition: all 0.2s ease;
-  box-shadow: 0 1px 2px rgba(0, 0, 0, 0.1);
-}
-
-.githubLink:hover {
-  background-color: #24292e;
-  border-color: #24292e;
-  color: #ffffff;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.2);
-  transform: translateY(-1px);
-}
-
-.githubLink:hover svg {
-  color: #ffffff;
-}
-
-.githubLink svg {
-  font-size: 18px;
-  color: #24292e;
-}
-
-.homeLink {
-  display: inline-flex;
-  align-items: center;
-  gap: 8px;
-  padding: 10px 16px;
-  background-color: #ffffff;
-  border: 1px solid #e5e7eb;
-  border-radius: 6px;
-  color: #374151;
-  text-decoration: none;
-  font-size: 14px;
-  font-weight: 600;
-  transition: all 0.2s ease;
-  box-shadow: 0 1px 2px rgba(0, 0, 0, 0.1);
-}
-
-.homeLink:hover {
-  background-color: #f9fafb;
-  border-color: #d1d5db;
-  transform: translateY(-1px);
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-.homeLink svg {
-  font-size: 18px;
-  color: #4f46e5;
-  transition: transform 0.2s ease;
-}
-
-.homeLink:hover svg {
-  transform: translateX(-2px);
-}
-
-.content {
-  display: flex;
-  flex-direction: column;
-  gap: 3rem;
-}
-
-.section {
-  display: flex;
-  flex-direction: column;
-  gap: 1.5rem;
-}
-
-.sectionTitle {
-  font-size: 1.5rem;
-  font-weight: bold;
-  color: #1f2937;
-}
-
-.text {
-  font-size: 1rem;
-  color: #4b5563;
-  line-height: 1.625;
-}
-
-.featureList {
-  display: grid;
-  gap: 2rem;
-  padding: 0;
-  list-style: none;
-}
-
-@media (min-width: 1024px) {
-  .featureList {
-    grid-template-columns: repeat(2, 1fr);
-  }
-}
-
-.featureItem {
-  display: flex;
-  gap: 1rem;
-  padding: 1.5rem;
-  background-color: white;
-  border-radius: 0.5rem;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
-  transition: transform 0.2s, box-shadow 0.2s;
-}
-
-.featureItem:hover {
-  transform: translateY(-2px);
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-}
-
-.featureIcon {
-  font-size: 1.5rem;
-  width: 3rem;
-  height: 3rem;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background-color: #f3f4f6;
-  border-radius: 0.5rem;
-  flex-shrink: 0;
-}
-
-.featureTitle {
-  font-size: 1.125rem;
-  font-weight: 600;
-  color: #1f2937;
-  margin-bottom: 0.5rem;
-}
-
-.featureText {
-  font-size: 0.875rem;
-  color: #6b7280;
-  line-height: 1.5;
-}
-
-.contactInfo {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-  padding: 1.5rem;
-  background-color: white;
-  border-radius: 0.5rem;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
-}
-
-.contactInfo p {
-  font-size: 0.875rem;
-  color: #4b5563;
-}
-
-.contactInfo strong {
-  color: #1f2937;
-  font-weight: 500;
-  margin-right: 0.5rem;
-}
-
-.logoText {
-  font-size: 2.5rem;
-  font-weight: bold;
-  background: linear-gradient(to right, #2563eb, #60a5fa);
-  -webkit-background-clip: text;
-  background-clip: text;
-  color: transparent;
-  transition: all 0.3s;
-}
-
-.logoText:hover {
-  background: linear-gradient(to right, #3b82f6, #93c5fd);
-  -webkit-background-clip: text;
-  background-clip: text;
-}
-.aboutLink {
-  position: absolute;
-  top: 1rem;
-  left: 1rem;
-}
-
-@media (max-width: 768px) {
-  .aboutLink {
-    top: 0.75rem;
-    left: 0.75rem;
-  }
-
-  .aboutLink a {
-    padding: 0.5rem;
-  }
-
-  .aboutLink span {
-    display: none;
-  }
-
-  .aboutLink svg {
-    font-size: 1.5rem;
-    margin: 0;
-  }
-}
-
-.aboutLink a {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  padding: 0.75rem 1.25rem;
-  background-color: white;
-  border: 1px solid #e5e7eb;
-  border-radius: 8px;
-  color: #374151;
-  text-decoration: none;
-  font-size: 0.875rem;
-  font-weight: 500;
-  transition: all 0.2s ease;
-  box-shadow: 0 1px 2px rgba(0, 0, 0, 0.05);
-}
-
-.aboutLink a:hover {
-  background-color: #f9fafb;
-  border-color: #d1d5db;
-  transform: translateY(-1px);
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-.aboutLink svg {
-  font-size: 1.25rem;
-  color: #4f46e5;
-  transition: transform 0.2s ease;
-}
-
-.aboutLink a:hover svg {
-  transform: translateX(-2px);
-}
-
-.aboutLink span {
-  font-size: 1.125rem;
-  line-height: 1;
-}
