--- frontend/src/pages/Login/Login.module.css
@@ -1,175 +1,0 @@
-.container {
-  min-height: 100vh;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 1rem;
-  background-color: #f3f4f6;
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-}
-
-.card {
-  width: 100%;
-  max-width: 28rem;
-  background-color: white;
-  border-radius: 0.75rem;
-  box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1), 0 2px 4px -1px rgba(0, 0, 0, 0.06);
-  padding: 2.5rem;
-}
-
-.header {
-  text-align: center;
-  margin-bottom: 2.5rem;
-}
-
-.logo {
-  font-size: 2.25rem;
-  font-weight: bold;
-  background: linear-gradient(to right, #2563eb, #60a5fa);
-  -webkit-background-clip: text;
-  background-clip: text;
-  color: transparent;
-  margin-bottom: 1.25rem;
-}
-
-.title {
-  font-size: 1.875rem;
-  font-weight: bold;
-  color: #1f2937;
-  margin-bottom: 0.75rem;
-}
-
-.subtitle {
-  color: #6b7280;
-  font-size: 1rem;
-}
-
-.form {
-  display: flex;
-  flex-direction: column;
-  gap: 1.5rem;
-}
-
-.formGroup {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-  text-align: left;
-}
-
-.label {
-  font-size: 0.875rem;
-  font-weight: 500;
-  color: #374151;
-  text-align: left;
-}
-
-.input {
-  width: 100%;
-  padding: 0.75rem 1rem;
-  border: 1px solid #d1d5db;
-  border-radius: 0.5rem;
-  font-size: 1rem;
-  color: #1f2937;
-  transition: all 0.2s;
-}
-
-.input:focus {
-  outline: none;
-  border-color: #3b82f6;
-  box-shadow: 0 0 0 3px rgba(59, 130, 246, 0.1);
-}
-
-.error {
-  color: #dc2626;
-  font-size: 0.875rem;
-  margin-top: 0.25rem;
-  text-align: left;
-}
-
-.button {
-  width: 100%;
-  padding: 0.875rem;
-  background-color: #3b82f6;
-  color: white;
-  font-weight: 500;
-  font-size: 1rem;
-  border-radius: 0.5rem;
-  transition: background-color 0.2s;
-}
-
-.button:hover {
-  background-color: #2563eb;
-}
-
-.button:disabled {
-  background-color: #9ca3af;
-  cursor: not-allowed;
-}
-
-.divider {
-  display: flex;
-  align-items: center;
-  margin: 1.5rem 0;
-  color: #6b7280;
-  font-size: 0.875rem;
-}
-
-.divider::before,
-.divider::after {
-  content: '';
-  flex: 1;
-  border-top: 1px solid #e5e7eb;
-}
-
-.divider span {
-  padding: 0 1rem;
-}
-
-.socialButtons {
-  display: flex;
-  gap: 1rem;
-}
-
-.socialButton {
-  flex: 1;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 0.5rem;
-  padding: 0.75rem;
-  border: 1px solid #d1d5db;
-  border-radius: 0.5rem;
-  font-size: 0.875rem;
-  color: #374151;
-  background-color: white;
-  transition: all 0.2s;
-}
-
-.socialButton:hover {
-  background-color: #f3f4f6;
-  border-color: #9ca3af;
-}
-
-.footer {
-  text-align: center;
-  margin-top: 1.5rem;
-  font-size: 0.875rem;
-  color: #6b7280;
-}
-
-.link {
-  color: #3b82f6;
-  font-weight: 500;
-  text-decoration: none;
-  transition: color 0.2s;
-}
-
-.link:hover {
-  color: #2563eb;
-  text-decoration: underline;
-}
