+++ frontend/src/pages/Login/index.tsx
@@ -1,6 +1,5 @@
 import React, { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
-import styles from './Login.module.css';
 import Loading from '../../components/Loading';
 import { authApi } from '../../api';
 import { Github, Mail } from 'lucide-react';
@@ -103,23 +102,23 @@ const Login: React.FC = () => {
   }, []);
 
   return (
-    <div className={styles.container}>
-      <div className={styles.card}>
-        <div className={styles.header}>
-          <div className={styles.logo}>TheLab</div>
-          <h2 className={styles.title}>欢迎回来</h2>
-          <p className={styles.subtitle}>请登录您的账号</p>
+    <div className="min-h-screen flex items-center justify-center p-4 bg-gray-100 fixed inset-0">
+      <div className="w-full max-w-md bg-white rounded-xl shadow-md p-10">
+        <div className="text-center mb-10">
+          <div className="text-4xl font-bold bg-gradient-to-r from-blue-600 to-blue-400 bg-clip-text text-transparent mb-5">TheLab</div>
+          <h2 className="text-3xl font-bold text-gray-800 mb-3">欢迎回来</h2>
+          <p className="text-gray-500">请登录您的账号</p>
         </div>
 
-        <form onSubmit={handleSubmit} className={styles.form}>
+        <form onSubmit={handleSubmit} className="flex flex-col gap-6">
           {errors.general && (
-            <div className={styles.errorMessage}>
+            <div className="text-red-600 text-sm">
               {errors.general}
             </div>
           )}
 
-          <div className={styles.formGroup}>
-            <label htmlFor="email" className={styles.label}>
+          <div className="flex flex-col gap-2 text-left">
+            <label htmlFor="email" className="text-sm font-medium text-gray-700">
               邮箱
             </label>
             <input
@@ -128,14 +127,14 @@ const Login: React.FC = () => {
               name="email"
               value={formData.email}
               onChange={handleChange}
-              className={styles.input}
+              className="w-full py-3 px-4 border border-gray-300 rounded-lg text-base text-gray-800 transition-all focus:outline-none focus:border-blue-500 focus:ring-2 focus:ring-blue-500/10"
               placeholder="your@email.com"
             />
-            {errors.email && <p className={styles.error}>{errors.email}</p>}
+            {errors.email && <p className="text-red-600 text-sm mt-1">{errors.email}</p>}
           </div>
 
-          <div className={styles.formGroup}>
-            <label htmlFor="password" className={styles.label}>
+          <div className="flex flex-col gap-2 text-left">
+            <label htmlFor="password" className="text-sm font-medium text-gray-700">
               密码
             </label>
             <input
@@ -144,30 +143,32 @@ const Login: React.FC = () => {
               name="password"
               value={formData.password}
               onChange={handleChange}
-              className={styles.input}
+              className="w-full py-3 px-4 border border-gray-300 rounded-lg text-base text-gray-800 transition-all focus:outline-none focus:border-blue-500 focus:ring-2 focus:ring-blue-500/10"
               placeholder="••••••••"
             />
-            {errors.password && <p className={styles.error}>{errors.password}</p>}
+            {errors.password && <p className="text-red-600 text-sm mt-1">{errors.password}</p>}
           </div>
 
           <button
             type="submit"
-            className={styles.button}
+            className="w-full py-3.5 bg-blue-500 text-white font-medium text-base rounded-lg transition-colors hover:bg-blue-600 disabled:bg-gray-400 disabled:cursor-not-allowed"
             disabled={loading}
           >
             {loading ? <Loading size="small" text="" /> : '登录'}
           </button>
         </form>
 
-        <div className={styles.divider}>
-          <span>或</span>
+        <div className="flex items-center my-6 text-gray-500 text-sm">
+          <div className="flex-1 border-t border-gray-200"></div>
+          <span className="px-4">或</span>
+          <div className="flex-1 border-t border-gray-200"></div>
         </div>
 
-        <div className={styles.socialButtons}>
+        <div className="flex gap-4">
           {providers.map(provider => (
             <button
               key={provider.name}
-              className={styles.socialButton}
+              className="flex-1 flex items-center justify-center gap-2 py-3 border border-gray-300 rounded-lg text-sm text-gray-700 bg-white hover:bg-gray-100 hover:border-gray-400 transition-all"
               onClick={() => window.location.href = provider.url}
             >
               <span className="flex items-center gap-2">{ICON_MAP[provider.name as keyof typeof ICON_MAP]} {provider.name} 登录</span>
@@ -175,9 +176,9 @@ const Login: React.FC = () => {
           ))}
         </div>
 
-        <div className={styles.footer}>
+        <div className="text-center mt-6 text-sm text-gray-500">
           还没有账号？{' '}
-          <Link to="/register" className={styles.link}>
+          <Link to="/register" className="text-blue-500 font-medium hover:text-blue-600 hover:underline">
             立即注册
           </Link>
         </div>
@@ -187,3 +188,4 @@ const Login: React.FC = () => {
 };
 
 export default Login;
+
